+++ src/main/java/com/apps/pochak/alarm/service/AlarmService.java
@@ -4,7 +4,7 @@ import com.apps.pochak.alarm.domain.Alarm;
 import com.apps.pochak.alarm.domain.repository.AlarmRepository;
 import com.apps.pochak.alarm.dto.response.AlarmElements;
 import com.apps.pochak.global.api_payload.code.BaseCode;
-import com.apps.pochak.login.jwt.JwtService;
+import com.apps.pochak.login.provider.JwtProvider;
 import com.apps.pochak.member.domain.Member;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
@@ -19,17 +19,17 @@ import static com.apps.pochak.global.api_payload.code.status.SuccessStatus.SUCCE
 @RequiredArgsConstructor
 public class AlarmService {
     private final AlarmRepository alarmRepository;
-    private final JwtService jwtService;
+    private final JwtProvider jwtProvider;
 
     @Transactional(readOnly = true)
     public AlarmElements getAllAlarms(Pageable pageable) {
-        final Member loginMember = jwtService.getLoginMember();
+        final Member loginMember = jwtProvider.getLoginMember();
         final Page<Alarm> alarmPage = alarmRepository.getAllAlarm(loginMember.getId(), pageable);
         return new AlarmElements(alarmPage);
     }
 
     public BaseCode checkAlarm(Long alarmId) {
-        final Member loginMember = jwtService.getLoginMember();
+        final Member loginMember = jwtProvider.getLoginMember();
         final Alarm alarm = alarmRepository.findAlarmById(alarmId, loginMember);
         alarm.setIsChecked(true);
         return SUCCESS_CHECK_ALARM;
