+++ src/main/java/com/apps/pochak/follow/service/FollowService.java
@@ -1,13 +1,12 @@
 package com.apps.pochak.follow.service;
 
-import com.apps.pochak.alarm.domain.Alarm;
 import com.apps.pochak.alarm.domain.FollowAlarm;
 import com.apps.pochak.alarm.domain.repository.AlarmRepository;
 import com.apps.pochak.follow.domain.Follow;
 import com.apps.pochak.follow.domain.repository.FollowRepository;
 import com.apps.pochak.global.api_payload.code.BaseCode;
 import com.apps.pochak.global.api_payload.exception.GeneralException;
-import com.apps.pochak.login.jwt.JwtService;
+import com.apps.pochak.login.provider.JwtProvider;
 import com.apps.pochak.member.domain.Member;
 import com.apps.pochak.member.domain.repository.CustomMemberRepository;
 import com.apps.pochak.member.domain.repository.MemberRepository;
@@ -19,7 +18,6 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.List;
 import java.util.Optional;
 
 import static com.apps.pochak.global.BaseEntityStatus.ACTIVE;
@@ -35,10 +33,10 @@ public class FollowService {
     private final AlarmRepository alarmRepository;
     private final MemberRepository memberRepository;
     private final CustomMemberRepository customMemberRepository;
-    private final JwtService jwtService;
+    private final JwtProvider jwtProvider;
 
     public BaseCode follow(final String handle) {
-        final Member loginMember = jwtService.getLoginMember();
+        final Member loginMember = jwtProvider.getLoginMember();
         final Member member = memberRepository.findByHandle(handle, loginMember);
 
         if (loginMember.getId().equals(member.getId())) {
@@ -95,7 +93,7 @@ public class FollowService {
                                    final String followerHandle
     ) {
         // TODO: Refactor permission checking part using annotations.
-        final Member loginMember = jwtService.getLoginMember();
+        final Member loginMember = jwtProvider.getLoginMember();
         if (!loginMember.getHandle().equals(handle)) {
             throw new GeneralException(_UNAUTHORIZED);
         }
@@ -113,7 +111,7 @@ public class FollowService {
     public MemberElements getFollowings(final String handle,
                                         final Pageable pageable
     ) {
-        final Member loginMember = jwtService.getLoginMember();
+        final Member loginMember = jwtProvider.getLoginMember();
         final Member member = memberRepository.findByHandle(handle, loginMember);
         final Page<MemberElement> followingPage = customMemberRepository.findFollowingsAndIsFollow(
                 member,
@@ -128,7 +126,7 @@ public class FollowService {
     public MemberElements getFollowers(final String handle,
                                        final Pageable pageable
     ) {
-        final Member loginMember = jwtService.getLoginMember();
+        final Member loginMember = jwtProvider.getLoginMember();
         final Member member = memberRepository.findByHandle(handle, loginMember);
         final Page<MemberElement> followerPage = customMemberRepository.findFollowersAndIsFollow(
                 member,
