+++ src/main/java/com/apps/pochak/report/service/ReportService.java
@@ -1,6 +1,6 @@
 package com.apps.pochak.report.service;
 
-import com.apps.pochak.login.jwt.JwtService;
+import com.apps.pochak.login.provider.JwtProvider;
 import com.apps.pochak.member.domain.Member;
 import com.apps.pochak.post.domain.Post;
 import com.apps.pochak.post.domain.repository.PostRepository;
@@ -17,10 +17,10 @@ import org.springframework.transaction.annotation.Transactional;
 public class ReportService {
     private final ReportRepository reportRepository;
     private final PostRepository postRepository;
-    private final JwtService jwtService;
+    private final JwtProvider jwtProvider;
 
     public void saveReport(final ReportUploadRequest request) {
-        final Member reporter = jwtService.getLoginMember();
+        final Member reporter = jwtProvider.getLoginMember();
         final Post reportedPost = postRepository.findPostById(request.getPostId(), reporter);
 
         Report report = request.toEntity(reporter, reportedPost);
